Move the setup dialog to the main window
No need for a separate file just for those few lines of code
